+++ lib/src/contained_tab_bar_view.dart
@@ -5,27 +5,58 @@ import 'enums.dart';
 import 'tab_bar_properties.dart';
 import 'tab_bar_view_properties.dart';
 
+/// ContainedTabBarView encapsulates [TabController], [TabBar] and [TabBarView]
+/// into a single, easy to use Widget.
+///
+/// It lets you customize its appearance without worrying
+/// about internal workings of "TabBar ecosystem".
 class ContainedTabBarView extends StatefulWidget {
-  final List<Widget> tabs;
-  final TabBarProperties tabBarProperties;
-  final List<Widget> views;
-  final TabBarViewProperties tabBarViewProperties;
-  final int initialIndex;
-  final void Function(int) onChange;
-  final bool callOnChangeWhileIndexIsChanging;
-
-  ContainedTabBarView({
-    Key key,
-    this.tabs,
-    this.tabBarProperties: const TabBarProperties(),
-    this.views,
-    this.tabBarViewProperties: const TabBarViewProperties(),
-    this.initialIndex: 0,
+  const ContainedTabBarView({
+    Key? key,
+    required this.tabs,
+    this.tabBarProperties = const TabBarProperties(),
+    required this.views,
+    this.tabBarViewProperties = const TabBarViewProperties(),
+    this.initialIndex = 0,
     this.onChange,
-    this.callOnChangeWhileIndexIsChanging: false,
-  })  : assert(tabs != null),
-        assert(views != null),
+    this.callOnChangeWhileIndexIsChanging = false,
+  })  : assert(
+          tabs.length == views.length,
+          'There has to be an equal amount of tabs (${tabs.length}) and views (${views.length}).',
+        ),
         super(key: key);
+
+  /// Typically a list of two or more widgets which are used as tab buttons.
+  ///
+  /// The amount of tabs has to match the amount of [views].
+  final List<Widget> tabs;
+
+  /// A [TabBarProperties] object which is used for customizing
+  /// the [TabBar]'s appearance.
+  final TabBarProperties tabBarProperties;
+
+  /// Typically a list of two or more widgets which are used as tab views.
+  ///
+  /// The amount of views has to match the amount of [tabs].
+  final List<Widget> views;
+
+  /// A [TabBarViewProperties] object which is used for customizing
+  /// the [TabBarView]'s appearance.
+  final TabBarViewProperties tabBarViewProperties;
+
+  /// The initial tab index.
+  final int initialIndex;
+
+  /// Everytime the tab index changed, this method is called.
+  ///
+  /// If [callOnChangeWhileIndexIsChanging] is true this method is
+  /// also called while we're animating from previousIndex to index
+  /// as a consequence of calling animateTo.
+  final void Function(int)? onChange;
+
+  /// Whether [onChange] should also get called while we're animating
+  /// from previousIndex to index as a consequence of calling animateTo.
+  final bool callOnChangeWhileIndexIsChanging;
 
   @override
   State<StatefulWidget> createState() => ContainedTabBarViewState();
@@ -33,7 +64,7 @@ class ContainedTabBarView extends StatefulWidget {
 
 class ContainedTabBarViewState extends State<ContainedTabBarView>
     with SingleTickerProviderStateMixin {
-  TabController _controller;
+  late TabController _controller;
 
   @override
   void initState() {
@@ -43,44 +74,44 @@ class ContainedTabBarViewState extends State<ContainedTabBarView>
       vsync: this,
       initialIndex: widget.initialIndex,
     )..addListener(() {
-        if (!widget.callOnChangeWhileIndexIsChanging) {
-          if (!_controller.indexIsChanging) {
-            widget.onChange(_controller.index);
-          }
-        } else {
-          widget.onChange(_controller.index);
+        if (widget.callOnChangeWhileIndexIsChanging ||
+            (!widget.callOnChangeWhileIndexIsChanging &&
+                !_controller.indexIsChanging)) {
+          widget.onChange?.call(_controller.index);
         }
       });
   }
 
-  // TODO expose it through onChange also
+  /// True while we're animating from previousIndex to index
+  /// as a consequence of calling animateTo.
   bool get indexIsChanging => _controller.indexIsChanging;
 
+  /// Animates to the tab with the given index.
   void animateTo(
     int value, {
-    Duration duration: kTabScrollDuration,
-    Curve curve: Curves.ease,
+    Duration duration = kTabScrollDuration,
+    Curve curve = Curves.ease,
   }) =>
       _controller.animateTo(value, duration: duration, curve: curve);
 
+  /// Animates to the next tab.
   void next({
-    Duration duration: kTabScrollDuration,
-    Curve curve: Curves.ease,
+    Duration duration = kTabScrollDuration,
+    Curve curve = Curves.ease,
   }) {
-    if (_controller.index == _controller.length - 1) {
-      return;
-    }
-    this.animateTo(_controller.index + 1);
+    if (_controller.index == _controller.length - 1) return;
+
+    animateTo(_controller.index + 1);
   }
 
+  /// Animates to the previous tab.
   void previous({
-    Duration duration: kTabScrollDuration,
-    Curve curve: Curves.ease,
+    Duration duration = kTabScrollDuration,
+    Curve curve = Curves.ease,
   }) {
-    if (_controller.index == 0) {
-      return;
-    }
-    this.animateTo(_controller.index - 1);
+    if (_controller.index == 0) return;
+
+    animateTo(_controller.index - 1);
   }
 
   @override
@@ -92,45 +123,44 @@ class ContainedTabBarViewState extends State<ContainedTabBarView>
   }
 
   Widget _buildFlex(BoxConstraints constraints) {
-    if (widget.tabBarProperties.position == TabBarPosition.left ||
-        widget.tabBarProperties.position == TabBarPosition.right) {
-      return Row(
-        crossAxisAlignment: _decideAlignment(widget.tabBarProperties.alignment),
-        children: _buildChildren(constraints),
-      );
-    }
-    return Column(
-      crossAxisAlignment: _decideAlignment(widget.tabBarProperties.alignment),
-      children: _buildChildren(constraints),
-    );
+    return widget.tabBarProperties.position == TabBarPosition.left ||
+            widget.tabBarProperties.position == TabBarPosition.right
+        ? Row(
+            crossAxisAlignment:
+                _decideAlignment(widget.tabBarProperties.alignment),
+            children: _buildChildren(constraints),
+          )
+        : Column(
+            crossAxisAlignment:
+                _decideAlignment(widget.tabBarProperties.alignment),
+            children: _buildChildren(constraints),
+          );
   }
 
   List<Widget> _buildChildren(BoxConstraints constraints) {
     List<Widget> children = [
       Padding(
-        padding: widget.tabBarProperties.outerPadding,
+        padding: widget.tabBarProperties.margin,
         child: _buildTabBar(),
       ),
       _buildTabBarView(constraints),
     ];
 
-    if (widget.tabBarProperties.position == TabBarPosition.bottom ||
-        widget.tabBarProperties.position == TabBarPosition.right) {
-      return children.reversed.toList();
-    }
-
-    return children;
+    return widget.tabBarProperties.position == TabBarPosition.bottom ||
+            widget.tabBarProperties.position == TabBarPosition.right
+        ? children.reversed.toList()
+        : children;
   }
 
   Widget _buildTabBar() {
     final List<Widget> backgroundStackChildren = [];
     if (widget.tabBarProperties.background != null) {
-      backgroundStackChildren.add(widget.tabBarProperties.background);
+      backgroundStackChildren.add(widget.tabBarProperties.background!);
     }
     backgroundStackChildren.add(
       Positioned.fill(
         child: Padding(
-          padding: widget.tabBarProperties.innerPadding,
+          padding: widget.tabBarProperties.padding,
           child: TabBar(
             controller: _controller,
             tabs: widget.tabs,
@@ -159,57 +189,56 @@ class ContainedTabBarViewState extends State<ContainedTabBarView>
       ),
     );
 
-    if (widget.tabBarProperties.position == TabBarPosition.left) {
-      return Expanded(
-        child: Transform.rotate(
-          angle: -math.pi / 2,
-          child: tabBar,
-        ),
-      );
+    switch (widget.tabBarProperties.position) {
+      case TabBarPosition.left:
+        return Expanded(
+          child: Transform.rotate(
+            angle: -math.pi / 2,
+            child: tabBar,
+          ),
+        );
+      case TabBarPosition.right:
+        return Expanded(
+          child: Transform.rotate(
+            angle: math.pi / 2,
+            child: tabBar,
+          ),
+        );
+      default:
+        return tabBar;
     }
-
-    if (widget.tabBarProperties.position == TabBarPosition.right) {
-      return Expanded(
-        child: Transform.rotate(
-          angle: math.pi / 2,
-          child: tabBar,
-        ),
-      );
-    }
-
-    return tabBar;
   }
 
   Widget _buildTabBarView(BoxConstraints constraints) {
-    final EdgeInsets outerPadding =
-        widget.tabBarProperties.outerPadding.resolve(TextDirection.ltr);
-    if (widget.tabBarProperties.position == TabBarPosition.left ||
-        widget.tabBarProperties.position == TabBarPosition.right) {
-      return Container(
-        width: constraints.maxWidth -
-            widget.tabBarProperties.height -
-            outerPadding.top -
-            outerPadding.bottom,
-        child: TabBarView(
-          controller: _controller,
-          children: widget.views,
-          dragStartBehavior: widget.tabBarViewProperties.dragStartBehavior,
-          physics: widget.tabBarViewProperties.physics,
-        ),
-      );
-    }
-    return Container(
-      height: constraints.maxHeight -
-          widget.tabBarProperties.height -
-          outerPadding.top -
-          outerPadding.bottom,
-      child: TabBarView(
-        controller: _controller,
-        children: widget.views,
-        dragStartBehavior: widget.tabBarViewProperties.dragStartBehavior,
-        physics: widget.tabBarViewProperties.physics,
-      ),
-    );
+    final EdgeInsets margin =
+        widget.tabBarProperties.margin.resolve(TextDirection.ltr);
+
+    return widget.tabBarProperties.position == TabBarPosition.left ||
+            widget.tabBarProperties.position == TabBarPosition.right
+        ? Container(
+            width: constraints.maxWidth -
+                widget.tabBarProperties.height -
+                margin.top -
+                margin.bottom,
+            child: TabBarView(
+              controller: _controller,
+              children: widget.views,
+              dragStartBehavior: widget.tabBarViewProperties.dragStartBehavior,
+              physics: widget.tabBarViewProperties.physics,
+            ),
+          )
+        : Container(
+            height: constraints.maxHeight -
+                widget.tabBarProperties.height -
+                margin.top -
+                margin.bottom,
+            child: TabBarView(
+              controller: _controller,
+              children: widget.views,
+              dragStartBehavior: widget.tabBarViewProperties.dragStartBehavior,
+              physics: widget.tabBarViewProperties.physics,
+            ),
+          );
   }
 
   CrossAxisAlignment _decideAlignment(TabBarAlignment alignment) {